+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- MV-339 - Display call visio button in the networking private chat
+
 ## [2.94.1] - 2021-06-03
 ### Fixed
 - MV-338 - Hide meeting request on sheet if not allowed by seewhat rule